+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.0 (????-??-??)
 
-- Requires Java 21 to run.
+- Requires Java 17 to run.
 
 ## 0.9.0 (2025-12-10)